feat(MFLP-73): Separate order domain modules
