+++ buyhood-order/build.gradle
@@ -1,0 +1,65 @@
+plugins {
+	id 'java'
+	id 'org.springframework.boot' version '3.5.0'
+	id 'io.spring.dependency-management' version '1.1.7'
+}
+
+group = 'api'
+version = '0.0.1-SNAPSHOT'
+
+ext {
+	set('springCloudVersion', "2025.0.0")
+}
+
+dependencyManagement {
+	imports {
+		mavenBom "org.springframework.cloud:spring-cloud-dependencies:${springCloudVersion}"
+	}
+}
+
+
+java {
+	toolchain {
+		languageVersion = JavaLanguageVersion.of(21)
+	}
+}
+
+repositories {
+	mavenCentral()
+	maven { url 'https://jitpack.io' }
+}
+
+tasks.named("bootJar") {
+	enabled = true
+}
+
+dependencies {
+	implementation project(':buyhood-global-core:buyhood-global-config')
+	implementation project(':buyhood-user')
+	implementation project(':buyhood-cart')
+	implementation project(':buyhood-product')
+	implementation project(':buyhood-store')
+	implementation project(':buyhood-payment')
+
+	// spring validation
+	implementation 'org.springframework.boot:spring-boot-starter-validation'
+
+	// h2 database
+	runtimeOnly 'com.h2database:h2'
+
+	// mysql driver
+	runtimeOnly 'com.mysql:mysql-connector-j'
+
+	//portone
+	implementation 'com.github.iamport:iamport-rest-client-java:0.2.23'
+	implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+
+}
+
+tasks.named('test') {
+	useJUnitPlatform()
+}
+
+tasks.register("prepareKotlinBuildScriptModel") {
+
+}
