+++ buyhood-order/src/main/java/api/buyhood/order/entity/OrderHistory.java
@@ -1,0 +1,40 @@
+package api.buyhood.order.entity;
+
+import api.buyhood.entity.BaseTimeEntity;
+import api.buyhood.product.entity.Product;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Table(name = "order_histories")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class OrderHistory extends BaseTimeEntity {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	@Column(nullable = false)
+	private Long id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "order_id")
+	private Order order;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "product_id")
+	private Product product;
+
+	@Column(nullable = false)
+	private int quantity;
+
+	@Builder
+	public OrderHistory(Order order, Product product, int quantity) {
+		this.order = order;
+		this.product = product;
+		this.quantity = quantity;
+	}
+
+}
