+++ buyhood-order/src/main/java/api/buyhood/order/enums/PaymentMethod.java
@@ -1,0 +1,12 @@
+package api.buyhood.order.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum PaymentMethod {
+    CARD,
+    VBANK,
+    TRANS,
+    ZERO_PAY,
+    PHONE;
+}
