+++ buyhood-order/src/test/java/api/buyhood/order/BuyHoodOrderApplicationTests.java
@@ -1,0 +1,13 @@
+package api.buyhood.order;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest(classes = BuyHoodOrderApplication.class)
+class BuyHoodOrderApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
