+++ settings.gradle
@@ -8,5 +8,8 @@ include 'buyhood-global-core:buyhood-global-feign'
 include 'buyhood-category'
 include 'buyhood-category:buyhood-store-category'
 include 'buyhood-category:buyhood-product-category'
+include 'buyhood-order'
+include 'buyhood-cart'
 include 'buyhood-product'
 include 'buyhood-store'
+include 'buyhood-payment'
