--- src/main/java/api/buyhood/domain/order/entity/OrderHistory.java
@@ -1,48 +1,0 @@
-package api.buyhood.domain.order.entity;
-
-import api.buyhood.domain.product.entity.Product;
-import api.buyhood.entity.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@Table(name = "order_histories")
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class OrderHistory extends BaseTimeEntity {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	@Column(nullable = false)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "order_id")
-	private Order order;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "product_id")
-	private Product product;
-
-	@Column(nullable = false)
-	private int quantity;
-
-	@Builder
-	public OrderHistory(Order order, Product product, int quantity) {
-		this.order = order;
-		this.product = product;
-		this.quantity = quantity;
-	}
-
-}
