--- src/main/java/api/buyhood/domain/order/enums/OrderStatus.java
@@ -1,16 +1,0 @@
-package api.buyhood.domain.order.enums;
-
-public enum OrderStatus {
-	//공용
-	PENDING,
-	CANCELED,
-	ACCEPTED,
-	REJECTED,
-	COMPLETED,
-
-	//배달가능
-	DELIVERING,
-	DELIVERED,
-	//ONLY 픽업
-	READY
-}
