--- src/main/java/api/buyhood/domain/order/enums/PaymentMethod.java
@@ -1,12 +1,0 @@
-package api.buyhood.domain.order.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum PaymentMethod {
-    CARD,
-    VBANK,
-    TRANS,
-    ZERO_PAY,
-    PHONE;
-}
